refactor(d3): modernize codebase and update tooling config
- Add clangd configuration and compatibility shim.
- Update clang-tidy rules and apply fixes across the codebase.
- Rename global variables and constants to standard prefixes (`g_`, `k`, `s_`).
- Improve type safety with explicit casts and `nullptr` usage.
- Update `FuncPtrTraits` to support `noexcept` function pointers.
- Refactor NVN hooks logic into `d3::nvn` namespace.
- Reduce default memory pool sizes in `setting.hpp`.
- Suppress warnings for third-party includes in makefile.
- Move global variable definitions from `_util.hpp` to new `_util_globals.cpp` to resolve ODR violations and prevent multiple definitions.
- Update header declarations to use `extern` and `inline` appropriately.
- Refactor `AllGBIDsOfType` to accept vectors by reference instead of pointer.
- Modernize C++ syntax with `using` aliases and explicit `std::` namespace qualifiers.
- Update CMake include directories and clang-tidy configuration.
- Optimize ImGui backend configuration by disabling unused features.
- Update CMake toolchain to detect GCC sysroot and include paths.
- Configure .clangd and .clang-tidy for C++23 and target flags.
- Adjust type definitions in idadefs.h and NEON types for Clang.
- Remove redundant alloc.hpp content and fix lobby.hpp includes.

+++ source/program/d3/types/refstring.cpp
@@ -1,5 +1,4 @@
 #include "symbols/refstring.hpp"
-#include "d3/setting.hpp"
 
 SETUP_CRS_PTR(CRefString_ctor_default);
 SETUP_CRS_PTR(CRefString_ctor_int);
@@ -51,7 +50,7 @@ CRefString::CRefString(const char *szString) {
     CRefString_ctor_lpcstr(this, szString);
 }
 
-CRefString::CRefString(const char *szString, int nLength) { CRefString_CommonCtorBody(this, szString, static_cast<int>(nLength)); }
+CRefString::CRefString(const char *szString, int nLength) { CRefString_CommonCtorBody(this, szString, nLength); }
 CRefString::CRefString(const char *szString, uint nLength) { CRefString_CommonCtorBody(this, szString, static_cast<int>(nLength)); }
 CRefString::CRefString(const char *szString, long nLength) { CRefString_CommonCtorBody(this, szString, static_cast<int>(nLength)); }
